Nemmando Single Tap changes ( + a util)
Single Tap now has a new keyword!
-------------------------------------------
Rending makes it so upon hitting an enemy with gouge, the damage of the hit is doubled and a stack of gouge is consumed.
-------------------------------------------
This aims to make nemmando's alt secondary much more interesting by giving it unqiue interactions, while also adding more mechanical depth to nemmando. Hopefully you guys enjoy!

Also, added a new util that should make it much easier to modify the stack count of Dots